+++ src/components/utilities/Defense/Overview/Table/DefenseOverviewTable.vue
@@ -9,22 +9,30 @@
     <table class="table table-striped table-sm position-relative" :class="{ 'table-hover': tableHover }">
       <LoadingSpinner v-if="loading" />
       <thead>
-        <tr>
+        <tr ref="tableHeaderWrapper">
+          <th scope="col">
+            <div class="dropdown">
+              <IconElipsis class="btn--context-menu" data-bs-toggle="dropdown" data-bs-auto-close="outside" aria-expanded="false" />
+              <ul class="dropdown-menu">
+                <li v-for="availableTableHeader in availableTableHeaders" :key="availableTableHeader.key">
+                  <span class="dropdown-item" @click.prevent="tableHeaders[availableTableHeader.key].visible = !tableHeaders[availableTableHeader.key].visible">
+                    <IconCheck class="column-visibility-indicator column-visibility-indicator--visible" v-if="tableHeaders[availableTableHeader.key].visible" />
+                    <IconCross class="column-visibility-indicator column-visibility-indicator--hidden" v-else />
+                    {{ availableTableHeader.label }}
+                  </span>
+                </li>
+              </ul>
+            </div>
+          </th>
           <th scope="col"><input type="checkbox" :checked="allChecked" @change="checkAll" /></th>
-          <th scope="col"></th>
           <th scope="col">Icon</th>
           <th scope="col">Label</th>
-          <th scope="col">Hit Points</th>
-          <th scope="col">Rate</th>
-          <th scope="col">Range</th>
-          <th scope="col">Crit. Chance</th>
-          <th scope="col">Crit. Damage</th>
-          <th scope="col">Tooltip DPS</th>
-          <th scope="col">Actual DPS</th>
+          <th class="column--dragable" v-for="tableHeader in tableHeaders" v-show="tableHeader.visible" :key="tableHeader.key" scope="col">{{ tableHeader.label }}</th>
           <th scope="col" class="text-center">Tier #</th>
+          <th scope="col"></th>
         </tr>
       </thead>
-      <tbody>
+      <tbody ref="tableRowsWrapper">
         <slot name="defense-list" v-for="defense in defenses" :defense="defense" :allChecked="allChecked" :selectDefenseCallback="selectDefense" :key="defense.incrementId">
           <Defense
             :all-checked="allChecked"
@@ -48,16 +56,28 @@
 <script setup lang="ts">
 import { ref, computed, defineProps, defineEmits, onMounted, toRef } from "vue";
 import { storeToRefs } from "pinia";
+import { useDraggable } from 'vue-draggable-plus'
 
 import { useGoogleSpreadsheetDataStore } from "@/stores/GoogleSpreadSheets";
 
 import type { ToRef } from "vue";
 import type { UserDataStoreDefenseInterface } from "@/stores/UserData";
 
+import { useUserDataStore } from "@/stores/UserData";
+
 import Defense from "@/components/utilities/Defense/Defense.vue";
 import LoadingSpinner from "@/components/layout/LoadingSpinner.vue";
 import BootstrapModal from "@/components/layout/BootstrapModal.vue";
 import DefenseUserInfo from "@/components/utilities/Defense/DefenseUserInfo.vue";
+import IconElipsis from '@/components/icons/IconElipsis.vue';
+import IconCheck from '@/components/icons/IconCheck.vue';
+import IconCross from '@/components/icons/IconCross.vue';
+
+export interface TableHeaderInterface {
+  label: string;
+  key: string;
+  visible: boolean;
+}
 
 const { loading } = storeToRefs(useGoogleSpreadsheetDataStore())
 const props = defineProps({
@@ -66,8 +86,27 @@ const props = defineProps({
 });
 
 const defenses: ToRef<UserDataStoreDefenseInterface[]> = toRef(props, 'defenses') as ToRef<UserDataStoreDefenseInterface[]>;
+const { tableHeaders } = storeToRefs(useUserDataStore())
 
 const emit = defineEmits(['totalDpsCalculated', 'deleteDefense']);
+
+const availableTableHeaders = ref<TableHeaderInterface[]>([
+  { key: "defenseHitPoints", label: "Hit Points", visible: true },
+  { key: "attackRate", label: "Rate", visible: true },
+  { key: "defenseRange", label: "Range", visible: true },
+  { key: "criticalChance", label: "Crit. Chance", visible: true },
+  { key: "criticalDamage", label: "Crit. Damage", visible: true },
+  { key: "tooltipDps", label: "Tooltip DPS", visible: true },
+  { key: "totalDps", label: "Actual DPS", visible: true },
+]);
+const tableHeaderWrapper = ref();
+const tableRowsWrapper = ref();
+
+// TODO Fix a few reactivity issues with sorting header columns
+/** @ts-ignore */
+useDraggable(tableHeaderWrapper, tableHeaders, { animation: 150, draggable: '.column--dragable' })
+/** @ts-ignore */
+useDraggable(tableRowsWrapper, props.defenses, { animation: 150, handle: '.sorting-handle' })
 
 const allChecked = ref<boolean>(false);
 const editDefense = ref<UserDataStoreDefenseInterface>();
@@ -137,6 +176,33 @@ th {
     display: flex;
     justify-content: flex-end;
     margin-top: 1rem;
+  }
+}
+
+.column--dragable {
+  cursor: move;
+}
+.btn--context-menu {
+  cursor: pointer;
+  height: 15px;
+  margin-bottom: 5px;
+}
+
+.dropdown-item {
+  display: flex;
+  align-items: center;
+  cursor: pointer;
+}
+.column-visibility-indicator {
+  height: 18px;
+  margin-right: 5px;
+
+  &--visible {
+    color: #056905;
+  }
+  
+  &--hidden {
+    color: #921515;
   }
 }
 </style>