+++ src/components/utilities/Defense/Overview/Table/DefenseOverviewTableRow.vue
@@ -1,10 +1,10 @@
 <template>
   <tr @click="$emit('defenseEdit')">
+    <!-- Handle for sorting -->
+    <td @click.stop><IconBars class="sorting-handle" /></td>
+
     <!-- Select for actions -->
     <td @click.stop><input type="checkbox" :checked="allChecked || checked" @change="({target}) => $emit('rowSelect', checked = (target as HTMLInputElement)?.checked)" /></td>
-
-    <!-- Damage type icon -->
-    <td @click.stop style="width: 30px"><DefenseDamageTypeIcon :defense="defense" /></td>
 
     <!-- Icon -->
     <td style="width: 50px"><img width="50" :src="icon" :alt="'Defense Icon ' + label"></td>
@@ -12,39 +12,13 @@
     <!-- Label -->
     <td>{{ label }}</td>
 
-    <!-- Defense HP -->
-    <td>{{ defense.isBuffDefense ? 'N/A' : Math.round(defenseStats.defenseHitPoints).toLocaleString('en-US') }}</td>
+    <td v-for="tableHeader in tableHeaders" v-show="tableHeader.visible" :key="tableHeader.key">
+      {{ statForHeader(tableHeader) }}
 
-    <!-- Defense Rate -->
-    <td v-if="!isBuffDefense">{{ defenseStats.attackRate.toFixed(3).replace(/(\.[^0]*)0+$/, '$1').replace(/\.$/, '') }} ({{ defenseStats.attackRatePercentage }}%)</td>
-    <td v-else>N/A</td>
-
-    <!-- Defense Range -->
-    <td>{{ defenseStats.defenseRange }}</td>
-
-    <!-- Critical chance -->
-    <td>{{ isBuffDefense ? 'N/A' : (defenseStats.criticalChance.toFixed(2) + '%') }}</td>
-
-    <!-- Critical damage -->
-    <td v-if="isBuffDefense">
-      {{ (defenseStats.criticalDamage / 4).toFixed(2) }}%
-      <HtmlTooltip>
-        This is the critical damage bonus applied to other defenses
+      <HtmlTooltip v-if="isBuffDefense && (tableHeader.key === 'criticalDamage' || tableHeader.key === 'tooltipDps')">
+        This is the bonus applied to other defenses
       </HtmlTooltip>
     </td>
-    <td v-else>{{ (defenseStats.criticalDamage).toFixed(2) }}%</td>
-
-    <!-- Tooltip DPS -->
-    <td v-if="isBuffDefense">
-      {{ Math.round(defenseStats.defensePower / 10).toLocaleString('en-US') }}
-      <HtmlTooltip>
-        This is the defense power bonus applied to other defenses
-      </HtmlTooltip>
-    </td>
-    <td v-else>{{ isBuffDefense ? 'N/A' : Math.round(defenseStats.tooltipDps).toLocaleString('en-US') }}</td>
-
-    <!-- Actual DPS -->
-    <td>{{ defenseStats.totalDps > 0 ? Math.round(defenseStats.totalDps).toLocaleString('en-US') : 'N/A' }}</td>
 
     <!-- Defense upgrade tier -->
     <td class="p-0" @click.stop>
@@ -54,22 +28,30 @@
         <button class="btn btn-link btn--down" @click="$emit('defenseLevelUpdate', defenseLevel-1)" :disabled="defenseLevel===1"><IconChevronDown /></button>
       </div>
     </td>
+
+    <!-- Damage type icon -->
+    <td @click.stop style="width: 30px"><DefenseDamageTypeIcon :defense="defense" /></td>
   </tr>
 </template>
 
 <script setup lang="ts">
 import { ref, defineProps } from "vue";
+import { storeToRefs } from "pinia";
 
 import type { PropType } from "vue";
 import type { DefenseStatsInterface } from "@/types";
+import type { UserDataStoreDefenseInterface } from "@/stores/UserData";
+import type { TableHeaderInterface } from "./DefenseOverviewTable.vue";
+
+import { useUserDataStore } from "@/stores/UserData";
 
 import HtmlTooltip from "@/components/layout/HtmlTooltip.vue";
-import type {UserDataStoreDefenseInterface} from "@/stores/UserData";
 import DefenseDamageTypeIcon from "@/components/utilities/Defense/DefenseDamageTypeIcon.vue";
 import IconChevronDown from "@/components/icons/IconChevronDown.vue";
 import IconChevronUp from "@/components/icons/IconChevronUp.vue";
+import IconBars from "@/components/icons/IconBars.vue";
 
-defineProps({
+const props = defineProps({
   icon: String,
   label: String,
   allChecked: Boolean,
@@ -89,7 +71,40 @@ defineProps({
   },
 });
 
+const { tableHeaders } = storeToRefs(useUserDataStore())
 const checked = ref<boolean>(false);
+
+function statForHeader(tableHeader: TableHeaderInterface): string | number {
+  let headerStat: number = (props.defenseStats as any)[tableHeader.key];
+  
+  if (props.defense.isBuffDefense) {
+    switch (tableHeader.key) {
+      case 'defenseHitPoints':
+      case 'attackRate':
+      case 'criticalChance':
+        return 'N/A';
+      case 'criticalDamage':
+        return (headerStat / 4).toFixed(2) + '%';
+      case 'tooltipDps':
+        headerStat = props.defenseStats.defensePower / 10;
+    }
+  }
+
+  switch (tableHeader.key) {
+    case 'attackRate':
+      return headerStat.toFixed(3).replace(/(\.[^0]*)0+$/, '$1').replace(/\.$/, '') + `(${props.defenseStats.attackRatePercentage}%)`
+    case 'criticalChance':
+    case 'criticalDamage':
+      return headerStat.toFixed(2) + '%';
+    case 'defenseHitPoints':
+    case 'tooltipDps':
+    case 'totalDps':
+      return headerStat > 0 ? Math.round(headerStat).toLocaleString('en-US') : ''
+    case 'defenseRange':
+    default:
+      return headerStat;
+  }
+}
 </script>
 
 <style lang="scss" scoped>
@@ -123,5 +138,11 @@ td {
       height: 13px;
     }
   }
+}
+
+.sorting-handle {
+  cursor: pointer;
+  height: 15px;
+  margin-bottom: 4px;
 }
 </style>