+++ src/stores/UserData.ts
@@ -8,6 +8,7 @@ import type {
     ShardInterface, ModInterface
 } from "@/types";
 import type { UserAncientResetPoints } from "@/data/AncientPowers";
+import type { TableHeaderInterface } from '@/components/utilities/Defense/Overview/Table/DefenseOverviewTable.vue';
 import DataMigrations from "@/data/DataMigrations";
 import { useDefenseStore } from "@/stores/DefenseInfo";
 import { useModStore } from "@/stores/ModInfo";
@@ -34,10 +35,10 @@ export const useUserDataStore = defineStore('userDataStore', () => {
 
     const dataMigration = new DataMigrations
 
-    const colorMode = ref<string>(localStorage.getItem('colorMode') ?? 'dark')
     const tableView = ref<boolean>(localStorage.getItem('tableView') === 'true')
     const lastVisitedVersion = ref<string>(localStorage.getItem('lastVisitedVersion') ?? 'v0.0.0')
     const isDev = ref<boolean>(localStorage.getItem('isDev') === 'true')
+    const tableHeaders = ref<{ [key: string]: TableHeaderInterface}>(getTableHeaders())
     const defenses = ref<UserDataStoreDefenseInterface[]>(getDefenses())
     const defenseSetups = ref<UserDefenseSetupInterface[]>(getDefenseSetups())
     const ancientPowerPoints = ref<UserAncientResetPoints>(getAncientPowerPoints())
@@ -60,6 +61,23 @@ export const useUserDataStore = defineStore('userDataStore', () => {
         }
 
         return allDefenses
+    }
+
+    function getTableHeaders(): { [key: string]: TableHeaderInterface} {
+        const tableHeaders: { [key: string]: TableHeaderInterface} = JSON.parse(localStorage.getItem('tableHeaders') ?? '{}')
+
+        if (Object.keys(tableHeaders).length > 0) return tableHeaders
+
+        const defaultTableHeaders: { [key: string]: TableHeaderInterface} = {
+            defenseHitPoints: { key: "defenseHitPoints", label: "Hit Points", visible: true },
+            attackRate: { key: "attackRate", label: "Rate", visible: true },
+            defenseRange: { key: "defenseRange", label: "Range", visible: true },
+            criticalChance: { key: "criticalChance", label: "Crit. Chance", visible: true },
+            criticalDamage: { key: "criticalDamage", label: "Crit. Damage", visible: true },
+            tooltipDps: { key: "tooltipDps", label: "Tooltip DPS", visible: true },
+            totalDps: { key: "totalDps", label: "Actual DPS", visible: true },
+        };
+        return defaultTableHeaders
     }
 
     function getDefenseSetups(): UserDefenseSetupInterface[] {
@@ -195,9 +213,9 @@ export const useUserDataStore = defineStore('userDataStore', () => {
         localStorage.setItem('ancientResetPoints', JSON.stringify(ancientPowerPoints.value))
     }, { deep: true })
 
-    watch(colorMode, () => {
-        localStorage.setItem('colorMode', colorMode.value)
-    })
+    watch(tableHeaders, () => {
+        localStorage.setItem('tableHeaders', JSON.stringify(tableHeaders.value))
+    }, { deep: true })
 
     watch(tableView, () => {
         localStorage.setItem('tableView', tableView.value ? 'true' : 'false')
@@ -207,7 +225,7 @@ export const useUserDataStore = defineStore('userDataStore', () => {
         localStorage.setItem('lastVisitedVersion', lastVisitedVersion.value)
     })
 
-    return { isDev, colorMode, tableView, lastVisitedVersion, defenses, defenseSetups, ancientPowerPoints, deleteDefense, deleteDefenseSetup, getNextDefenseIncrementId, getNextDefenseSetupIncrementId, importDefenses, importDefenseSetups }
+    return { isDev, tableHeaders, tableView, lastVisitedVersion, defenses, defenseSetups, ancientPowerPoints, deleteDefense, deleteDefenseSetup, getNextDefenseIncrementId, getNextDefenseSetupIncrementId, importDefenses, importDefenseSetups }
 })
 
 export function getDefaultSetupModifiers(): DefenseSetupModifiersInterface {